게시글과 댓글의 수정/삭제 권한 부여하기

+++ app/controllers/board_controller.rb
@@ -50,6 +50,10 @@ class BoardController < ApplicationController
 
   def edit
 		@post = Post.find(params[:id])
+		if @post.user_id != session[:user_id]
+			flash[:alert] = "수정 권한이 없습니다."
+			redirect_to :back
+		end
   end
 
   def edit_complete
@@ -57,7 +61,7 @@ class BoardController < ApplicationController
 		post.category = params[:post_category]
 		post.title = params[:post_title]
 		post.price = params[:post_price]
-#post.number = nil
+		post.number = parmas[:post_number]
 		post.content = params[:post_content]
 
 		if post.save
@@ -71,9 +75,15 @@ class BoardController < ApplicationController
 
   def delete_complete
 		post = Post.find(params[:id])
-		post.destroy
-		flash[:alert] = "삭제되었습니다."
-		redirect_to "/"
+	
+		if post.user_id == session[:user_id]
+			post.destroy
+			flash[:alert] = "삭제되었습니다."
+			redirect_to "/"
+		else
+			flash[:alert] = "삭제 권한이 없습니다."
+			redirect_to :back
+		end
   end
 
 	def write_comment_complete
@@ -89,8 +99,14 @@ class BoardController < ApplicationController
 
 	def delete_comment_complete
 		comment = Comment.find(params[:id])
-		comment.destroy
-		flash[:alert] = "댓글이 삭제되었습니다."
-		redirect_to "/board/show/#{comment.post_id}"
+
+		if comment.user_id == session[:user_id]
+			comment.destroy
+			flash[:alert] = "댓글이 삭제되었습니다."
+			redirect_to "/board/show/#{comment.post_id}"
+		else
+			flash[:alert] = "댓글 삭제 권한이 없습니다."
+			redirect_to :back
+		end
 	end
 end
